+++ common/miscres.cpp
@@ -907,7 +907,7 @@ top:
         {
             driver_put_string(s_row+1, 3, C_GENERAL_MED, "Item name:");
             driver_put_string(s_row+1, 16, C_GENERAL_HI, g_formula_name.c_str());
-            i = g_formula_name.length() + 1;
+            i = static_cast<int>(g_formula_name.length() + 1);
             driver_put_string(s_row+2, 3, C_GENERAL_MED, "Item file:");
             if (g_formula_filename.length() >= 29)
             {
